requirements.txt
argcomplete==1.1.1                                                                                     
Django==1.9.5                                                                                          
django-autoslug==1.9.3                                                                                 
django-cors-headers==1.1.0                                                                             
django-filter==0.12.0                                                                                  
django-model-utils==2.3.1                                                                              
django-mptt==0.8.3                                                                                     
django-notifications-hq==1.0                                                                           
django-reversion==1.10.1                                                                               
django-smart-selects==1.2.2                                                                            
djangorestframework==3.3.2                                                                             
drf-extensions==0.2.8                                                                                  
filemagic==1.6                                                                                         
gitdb==0.6.4                                                                                           
GitPython==1.0.2                                                                                       
image==1.5.3                                                                                           
jsonfield==1.0.3                                                                                       
logilab-common==1.2.0                                                                                  
Pillow==3.2.0                                                                                          
pytz==2015.7                                                                                           
six==1.10.0                                                                                            
smmap==0.9.0                                                                                           
thrift==0.10.0
Yapsy==1.11.223


-e git+https://github.com/peragro/peragro-at.git#egg=peragro-at
-e git+https://github.com/peragro/django-project.git#egg=django-project

